Reorganized integration with catch.
