+++ include/autocheck/reporter.hpp
@@ -142,7 +142,7 @@ AUTOCHECK_TOKENPASTE(checker,__LINE__).expect([](std::string const& msg){\
 
 #endif // ASSERT_TRUE
 
-#ifdef TEST_CASE
+#if defined(TEST_CASE) || defined(CATCH_TEST_CASE)
 
   class catch_reporter : public reporter {
     public:
@@ -156,57 +156,9 @@ AUTOCHECK_TOKENPASTE(checker,__LINE__).expect([](std::string const& msg){\
       virtual void failure(size_t tests, const char* reason) const {
         std::ostringstream out;
         report_failure(out, tests, reason);
-		FAIL(out.str());
+        FAIL(out.str());
       }
   };
-
-//  class msg_reporter : public reporter {
-//    public:
-//      msg_reporter(std::string& msg): msg(msg) {}
-//      std::string& msg;
-//      virtual void success(size_t tests, size_t max_tests,
-//          size_t trivial, distribution&& dist) const
-//      {
-//        report_success(std::clog, tests, max_tests, trivial, std::move(dist));
-//        msg = "";
-//      }
-//
-//      virtual void failure(size_t tests, const char* reason) const {
-//        std::ostringstream out;
-//        report_failure(out, tests, reason);
-//        msg = out.str();
-//      }
-//  };
-//
-//struct gtest_checker
-//{
-//    std::function<void(std::string const&)> expectfn;
-//    std::string msg;
-//    msg_reporter rep{msg};
-//    template<class FN>
-//    void operator <<(FN const& fn)
-//    {
-//        fn(rep);
-//        expectfn(msg);
-//    }
-//
-//    template<class FN>
-//    gtest_checker& expect(FN const& fn)
-//    {
-//        expectfn = fn;
-//        return *this;
-//    }
-//};
-//
-//#define AUTOCHECK_TOKENPASTE2(x, y) x ## y
-//#define AUTOCHECK_TOKENPASTE(x, y) AUTOCHECK_TOKENPASTE2(x, y)
-//
-//#define EXPECT_AUTOCHECK(_rep) autocheck::gtest_checker AUTOCHECK_TOKENPASTE(checker,__LINE__);\
-//AUTOCHECK_TOKENPASTE(checker,__LINE__).expect([](std::string const& msg){\
-//    bool const AUTOCHECK_SUCCESS = "" == msg;\
-//    EXPECT_TRUE(AUTOCHECK_SUCCESS) << msg;\
-//})\
-//<< [&](ac::msg_reporter const& _rep)
 
 #endif // TEST_CASE
